fix(room): align room/user/role id types with current DB schema

+++ src/main/java/DAJ2EE/demo/repository/RoomRepository.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RoomRepository extends JpaRepository<Room, Integer> {
+public interface RoomRepository extends JpaRepository<Room, Long> {
     List<Room> findAllByOrderByRoomNumberAsc();
 
     boolean existsByRoomNumberIgnoreCase(String roomNumber);
 
-    boolean existsByRoomNumberIgnoreCaseAndIdNot(String roomNumber, Integer id);
+    boolean existsByRoomNumberIgnoreCaseAndIdNot(String roomNumber, Long id);
 }
